refactor(stdlib): merge operations into existing modules per review feedback
Consolidate new operations into existing stdlib modules as requested:
- Merge string_ops.py into strings.py (regex, encoding, hashing ops)
- Merge json_ops.py into jsonlib.py (set, merge, filter, map, delete ops)
- Merge file_ops.py into io.py (list, glob, info, mkdir, copy, move, delete ops)
- Keep conditionals.py as new module (new category of operations)

All examples tested and working after refactoring.
