+++ runtime/stdlib/__init__.py
@@ -9,13 +9,10 @@ from . import iter as iterlib
 from . import llm as llm
 from . import tools as tools
 from . import conditionals as conditionals
-from . import json_ops as json_ops
-from . import string_ops as string_ops
-from . import file_ops as file_ops
 
 
 def register_all(ops_registry: dict, register_op):
     for mod in (math, strings, strings_extra, jsonlib, io, http, paths, iterlib, llm, tools, 
-                conditionals, json_ops, string_ops, file_ops):
+                conditionals):
         if hasattr(mod, "register"):
             mod.register(register_op)
